Initial commit
Created initial pages and data for the DB credentials, Functions, and added a creation file. I only made one table to test but we should make sure all others are added in this way programmatically. Then when we sync the code the DB will also sync up its format if we make changes.